backend/pyproject.toml
[project]
name = "solar-sim-api"
version = "0.1.0"
description = "Django backend for Solar-Sim"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.4",
    "psycopg[binary]>=3.2",
    "redis>=5.0",
    "httpx>=0.27",
    "stripe>=10.0",
    "gunicorn>=23.0",
    "whitenoise>=6.7",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.4",
    "pytest-django>=4.8",
    "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"
